+++ contracts/5/ConfigurableInterestBankConfig.sol
@@ -6,19 +6,19 @@ import './GoblinConfig.sol';
 
 interface InterestModel {
   /// @dev Return the interest rate per second, using 1e18 as denom.
-  function getInterestRate(uint256 debt, uint256 floating) external view returns (uint256);
+  function getInterestRate(uint debt, uint floating) external view returns (uint);
 }
 
 contract TripleSlopeModel {
-  using SafeMath for uint256;
+  using SafeMath for uint;
 
   /// @dev Return the interest rate per second, using 1e18 as denom.
-  function getInterestRate(uint256 debt, uint256 floating) external pure returns (uint256) {
-    uint256 total = debt.add(floating);
-    uint256 utilization = debt.mul(10000).div(total);
+  function getInterestRate(uint debt, uint floating) external pure returns (uint) {
+    uint total = debt.add(floating);
+    uint utilization = debt.mul(10000).div(total);
     if (utilization < 5000) {
       // Less than 50% utilization - 10% APY
-      return uint256(10e16) / 365 days;
+      return uint(10e16) / 365 days;
     } else if (utilization < 9500) {
       // Between 50% and 95% - 10%-25% APY
       return (10e16 + utilization.sub(5000).mul(15e16).div(10000)) / 365 days;
@@ -27,27 +27,27 @@ contract TripleSlopeModel {
       return (25e16 + utilization.sub(7500).mul(75e16).div(10000)) / 365 days;
     } else {
       // Not possible, but just in case - 100% APY
-      return uint256(100e16) / 365 days;
+      return uint(100e16) / 365 days;
     }
   }
 }
 
 contract ConfigurableInterestBankConfig is BankConfig, Ownable {
   /// The minimum ETH debt size per position.
-  uint256 public minDebtSize;
+  uint public minDebtSize;
   /// The portion of interests allocated to the reserve pool.
-  uint256 public getReservePoolBps;
+  uint public getReservePoolBps;
   /// The reward for successfully killing a position.
-  uint256 public getKillBps;
+  uint public getKillBps;
   /// Mapping for goblin address to its configuration.
   mapping(address => GoblinConfig) public goblins;
   /// Interest rate model
   InterestModel public interestModel;
 
   constructor(
-    uint256 _minDebtSize,
-    uint256 _reservePoolBps,
-    uint256 _killBps,
+    uint _minDebtSize,
+    uint _reservePoolBps,
+    uint _killBps,
     InterestModel _interestModel
   ) public {
     setParams(_minDebtSize, _reservePoolBps, _killBps, _interestModel);
@@ -59,9 +59,9 @@ contract ConfigurableInterestBankConfig is BankConfig, Ownable {
   /// @param _killBps The new reward for killing a position value.
   /// @param _interestModel The new interest rate model contract.
   function setParams(
-    uint256 _minDebtSize,
-    uint256 _reservePoolBps,
-    uint256 _killBps,
+    uint _minDebtSize,
+    uint _reservePoolBps,
+    uint _killBps,
     InterestModel _interestModel
   ) public onlyOwner {
     minDebtSize = _minDebtSize;
@@ -71,15 +71,18 @@ contract ConfigurableInterestBankConfig is BankConfig, Ownable {
   }
 
   /// @dev Set the configuration for the given goblins. Must only be called by the owner.
-  function setGoblins(address[] calldata addrs, GoblinConfig[] calldata configs) external onlyOwner {
+  function setGoblins(address[] calldata addrs, GoblinConfig[] calldata configs)
+    external
+    onlyOwner
+  {
     require(addrs.length == configs.length, 'bad length');
-    for (uint256 idx = 0; idx < addrs.length; idx++) {
+    for (uint idx = 0; idx < addrs.length; idx++) {
       goblins[addrs[idx]] = configs[idx];
     }
   }
 
   /// @dev Return the interest rate per second, using 1e18 as denom.
-  function getInterestRate(uint256 debt, uint256 floating) external view returns (uint256) {
+  function getInterestRate(uint debt, uint floating) external view returns (uint) {
     return interestModel.getInterestRate(debt, floating);
   }
 
@@ -94,12 +97,12 @@ contract ConfigurableInterestBankConfig is BankConfig, Ownable {
   }
 
   /// @dev Return the work factor for the goblin + ETH debt, using 1e4 as denom. Revert on non-goblin.
-  function workFactor(address goblin, uint256 debt) external view returns (uint256) {
+  function workFactor(address goblin, uint debt) external view returns (uint) {
     return goblins[goblin].workFactor(goblin, debt);
   }
 
   /// @dev Return the kill factor for the goblin + ETH debt, using 1e4 as denom. Revert on non-goblin.
-  function killFactor(address goblin, uint256 debt) external view returns (uint256) {
+  function killFactor(address goblin, uint debt) external view returns (uint) {
     return goblins[goblin].killFactor(goblin, debt);
   }
 }
