+++ contracts/5/Goblin.sol
@@ -3,9 +3,9 @@ pragma solidity 0.5.16;
 interface Goblin {
   /// @dev Work on a (potentially new) position. Optionally send ETH back to Bank.
   function work(
-    uint256 id,
+    uint id,
     address user,
-    uint256 debt,
+    uint debt,
     bytes calldata data
   ) external payable;
 
@@ -13,8 +13,8 @@ interface Goblin {
   function reinvest() external;
 
   /// @dev Return the amount of ETH wei to get back if we are to liquidate the position.
-  function health(uint256 id) external view returns (uint256);
+  function health(uint id) external view returns (uint);
 
   /// @dev Liquidate the given position to ETH. Send all ETH back to Bank.
-  function liquidate(uint256 id) external;
+  function liquidate(uint id) external;
 }
