+++ contracts/5/MStableGoblin.sol
@@ -16,13 +16,13 @@ import './Goblin.sol';
 contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   /// @notice Libraries
   using SafeToken for address;
-  using SafeMath for uint256;
+  using SafeMath for uint;
 
   /// @notice Events
-  event Reinvest(address indexed caller, uint256 reward, uint256 bounty);
-  event AddShare(uint256 indexed id, uint256 share);
-  event RemoveShare(uint256 indexed id, uint256 share);
-  event Liquidate(uint256 indexed id, uint256 wad);
+  event Reinvest(address indexed caller, uint reward, uint bounty);
+  event AddShare(uint indexed id, uint share);
+  event RemoveShare(uint indexed id, uint share);
+  event Liquidate(uint indexed id, uint wad);
 
   /// @notice Immutable variables
   IMStableStakingRewards public staking;
@@ -34,12 +34,12 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   address public operator;
 
   /// @notice Mutable state variables
-  mapping(uint256 => uint256) public shares;
+  mapping(uint => uint) public shares;
   mapping(address => bool) public okStrats;
-  uint256 public totalShare;
+  uint public totalShare;
   Strategy public addStrat;
   Strategy public liqStrat;
-  uint256 public reinvestBountyBps;
+  uint public reinvestBountyBps;
 
   constructor(
     address _operator,
@@ -48,7 +48,7 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
     address _mta,
     Strategy _addStrat,
     Strategy _liqStrat,
-    uint256 _reinvestBountyBps
+    uint _reinvestBountyBps
   ) public {
     operator = _operator;
     weth = _router.WETH();
@@ -62,9 +62,9 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
     okStrats[address(addStrat)] = true;
     okStrats[address(liqStrat)] = true;
     reinvestBountyBps = _reinvestBountyBps;
-    lpToken.approve(address(_staking), uint256(-1)); // 100% trust in the staking pool
-    lpToken.approve(address(router), uint256(-1)); // 100% trust in the router
-    _mta.safeApprove(address(router), uint256(-1)); // 100% trust in the router
+    lpToken.approve(address(_staking), uint(-1)); // 100% trust in the staking pool
+    lpToken.approve(address(router), uint(-1)); // 100% trust in the router
+    _mta.safeApprove(address(router), uint(-1)); // 100% trust in the router
   }
 
   /// @dev Require that the caller must be an EOA account to avoid flash loans.
@@ -81,17 +81,17 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
 
   /// @dev Return the entitied LP token balance for the given shares.
   /// @param share The number of shares to be converted to LP balance.
-  function shareToBalance(uint256 share) public view returns (uint256) {
+  function shareToBalance(uint share) public view returns (uint) {
     if (totalShare == 0) return share; // When there's no share, 1 share = 1 balance.
-    uint256 totalBalance = staking.balanceOf(address(this));
+    uint totalBalance = staking.balanceOf(address(this));
     return share.mul(totalBalance).div(totalShare);
   }
 
   /// @dev Return the number of shares to receive if staking the given LP tokens.
   /// @param balance the number of LP tokens to be converted to shares.
-  function balanceToShare(uint256 balance) public view returns (uint256) {
+  function balanceToShare(uint balance) public view returns (uint) {
     if (totalShare == 0) return balance; // When there's no share, 1 share = 1 balance.
-    uint256 totalBalance = staking.balanceOf(address(this));
+    uint totalBalance = staking.balanceOf(address(this));
     return balance.mul(totalShare).div(totalBalance);
   }
 
@@ -99,10 +99,10 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   function reinvest() public onlyEOA nonReentrant {
     // 1. Withdraw all the rewards.
     staking.claimReward();
-    uint256 reward = mta.myBalance();
+    uint reward = mta.myBalance();
     if (reward == 0) return;
     // 2. Send the reward bounty to the caller.
-    uint256 bounty = reward.mul(reinvestBountyBps) / 10000;
+    uint bounty = reward.mul(reinvestBountyBps) / 10000;
     mta.safeTransfer(msg.sender, bounty);
     // 3. Use add Two-side optimal strategy to convert MTA to ETH and add
     // liquidity to get LP tokens.
@@ -119,9 +119,9 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   /// @param debt The amount of user debt to help the strategy make decisions.
   /// @param data The encoded data, consisting of strategy address and calldata.
   function work(
-    uint256 id,
+    uint id,
     address user,
-    uint256 debt,
+    uint debt,
     bytes calldata data
   ) external payable onlyOperator nonReentrant {
     // 1. Convert this position back to LP tokens.
@@ -142,52 +142,52 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   /// @param rIn the amount of asset in reserve for input.
   /// @param rOut The amount of asset in reserve for output.
   function getMktSellAmount(
-    uint256 aIn,
-    uint256 rIn,
-    uint256 rOut
-  ) public pure returns (uint256) {
+    uint aIn,
+    uint rIn,
+    uint rOut
+  ) public pure returns (uint) {
     if (aIn == 0) return 0;
     require(rIn > 0 && rOut > 0, 'bad reserve values');
-    uint256 aInWithFee = aIn.mul(997);
-    uint256 numerator = aInWithFee.mul(rOut);
-    uint256 denominator = rIn.mul(1000).add(aInWithFee);
+    uint aInWithFee = aIn.mul(997);
+    uint numerator = aInWithFee.mul(rOut);
+    uint denominator = rIn.mul(1000).add(aInWithFee);
     return numerator / denominator;
   }
 
   /// @dev Return the amount of ETH to receive if we are to liquidate the given position.
   /// @param id The position ID to perform health check.
-  function health(uint256 id) external view returns (uint256) {
+  function health(uint id) external view returns (uint) {
     // 1. Get the position's LP balance and LP total supply.
-    uint256 lpBalance = shareToBalance(shares[id]);
-    uint256 lpSupply = lpToken.totalSupply(); // Ignore pending mintFee as it is insignificant
+    uint lpBalance = shareToBalance(shares[id]);
+    uint lpSupply = lpToken.totalSupply(); // Ignore pending mintFee as it is insignificant
     // 2. Get the pool's total supply of WETH and farming token.
-    (uint256 r0, uint256 r1, ) = lpToken.getReserves();
-    (uint256 totalWETH, uint256 totalMTA) = lpToken.token0() == weth ? (r0, r1) : (r1, r0);
+    (uint r0, uint r1, ) = lpToken.getReserves();
+    (uint totalWETH, uint totalMTA) = lpToken.token0() == weth ? (r0, r1) : (r1, r0);
     // 3. Convert the position's LP tokens to the underlying assets.
-    uint256 userWETH = lpBalance.mul(totalWETH).div(lpSupply);
-    uint256 userMTA = lpBalance.mul(totalMTA).div(lpSupply);
+    uint userWETH = lpBalance.mul(totalWETH).div(lpSupply);
+    uint userMTA = lpBalance.mul(totalMTA).div(lpSupply);
     // 4. Convert all farming tokens to ETH and return total ETH.
     return getMktSellAmount(userMTA, totalMTA.sub(userMTA), totalWETH.sub(userWETH)).add(userWETH);
   }
 
   /// @dev Liquidate the given position by converting it to ETH and return back to caller.
   /// @param id The position ID to perform liquidation
-  function liquidate(uint256 id) external onlyOperator nonReentrant {
+  function liquidate(uint id) external onlyOperator nonReentrant {
     // 1. Convert the position back to LP tokens and use liquidate strategy.
     _removeShare(id);
     lpToken.transfer(address(liqStrat), lpToken.balanceOf(address(this)));
     liqStrat.execute(address(0), 0, abi.encode(mta, 0));
     // 2. Return all available ETH back to the operator.
-    uint256 wad = address(this).balance;
+    uint wad = address(this).balance;
     SafeToken.safeTransferETH(msg.sender, wad);
     emit Liquidate(id, wad);
   }
 
   /// @dev Internal function to stake all outstanding LP tokens to the given position ID.
-  function _addShare(uint256 id) internal {
-    uint256 balance = lpToken.balanceOf(address(this));
+  function _addShare(uint id) internal {
+    uint balance = lpToken.balanceOf(address(this));
     if (balance > 0) {
-      uint256 share = balanceToShare(balance);
+      uint share = balanceToShare(balance);
       staking.stake(balance);
       shares[id] = shares[id].add(share);
       totalShare = totalShare.add(share);
@@ -196,10 +196,10 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   }
 
   /// @dev Internal function to remove shares of the ID and convert to outstanding LP tokens.
-  function _removeShare(uint256 id) internal {
-    uint256 share = shares[id];
+  function _removeShare(uint id) internal {
+    uint share = shares[id];
     if (share > 0) {
-      uint256 balance = shareToBalance(share);
+      uint balance = shareToBalance(share);
       staking.withdraw(balance);
       totalShare = totalShare.sub(share);
       shares[id] = 0;
@@ -214,14 +214,14 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   function recover(
     address token,
     address to,
-    uint256 value
+    uint value
   ) external onlyOwner nonReentrant {
     token.safeTransfer(to, value);
   }
 
   /// @dev Set the reward bounty for calling reinvest operations.
   /// @param _reinvestBountyBps The bounty value to update.
-  function setReinvestBountyBps(uint256 _reinvestBountyBps) external onlyOwner {
+  function setReinvestBountyBps(uint _reinvestBountyBps) external onlyOwner {
     reinvestBountyBps = _reinvestBountyBps;
   }
 
@@ -229,8 +229,8 @@ contract MStableGoblin is Ownable, ReentrancyGuard, Goblin {
   /// @param strats The strategy addresses.
   /// @param isOk Whether to approve or unapprove the given strategies.
   function setStrategyOk(address[] calldata strats, bool isOk) external onlyOwner {
-    uint256 len = strats.length;
-    for (uint256 idx = 0; idx < len; idx++) {
+    uint len = strats.length;
+    for (uint idx = 0; idx < len; idx++) {
       okStrats[strats[idx]] = isOk;
     }
   }
