+++ contracts/5/weth/IWETH.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.5.0;
 interface IWETH {
   function deposit() external payable;
 
-  function transfer(address to, uint256 value) external returns (bool);
+  function transfer(address to, uint value) external returns (bool);
 
-  function withdraw(uint256) external;
+  function withdraw(uint) external;
 }
